requirements.txt
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
deprecation==2.1.0
Django==5.0.3
django-background-tasks==1.2.5
django-compat==1.0.15
django-filter==24.1
django-storages==1.14.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gotrue==2.4.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
Markdown==3.6
packaging==24.0
pillow==10.2.0
postgrest==0.16.2
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
realtime==1.0.3
requests==2.31.0
rsa==4.9
schedule==1.2.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
storage3==0.7.4
StrEnum==0.4.15
stripe==5.5.0
supabase==2.4.1
supafunc==0.4.5
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
websockets==12.0
whitenoise==6.6.0
